Moving example to the same namespace so that it can use protected stuff
