+++ src/example/src/main/scala/gwi/mawex/ExampleExecutor.scala
@@ -1,12 +1,11 @@
-package example
+package gwi.mawex
 
 import akka.actor.{Actor, ActorLogging}
-import gwi.mawex._
 import gwi.mawex.Service.Address
 
 import scala.util.Success
 
-class Executor(args: Seq[String]) extends Actor with ActorLogging {
+class ExampleExecutor(args: Seq[String]) extends Actor with ActorLogging {
 
   // In real world, remote client would be a standalone service/container, here we start it from a Worker service to avoid building it
   Client.startRemoteClient(Address("workers", 0), List(Address("master-a", 2552), Address("master-b", 2551)))
