+++ src/example/src/main/scala/gwi/mawex/Producer.scala
@@ -1,9 +1,8 @@
-package example
+package gwi.mawex
 
 import java.util.UUID
 
 import akka.actor.{Actor, ActorLogging, ActorRef}
-import gwi.mawex._
 
 import scala.concurrent.duration._
 import scala.concurrent.forkjoin.ThreadLocalRandom
@@ -15,13 +14,12 @@ object Producer {
 class Producer(masterProxy: ActorRef) extends Actor with ActorLogging {
   import Producer._
   import context.dispatcher
-  val ConsumerGroup = "default"
-  val Pod = "default"
-  def scheduler = context.system.scheduler
-  def rnd = ThreadLocalRandom.current
-  def nextTaskId(): TaskId = TaskId(UUID.randomUUID().toString, ConsumerGroup)
+  private[this] val ConsumerGroup = "default"
+  private[this] def scheduler = context.system.scheduler
+  private[this] def rnd = ThreadLocalRandom.current
+  private[this] def nextTaskId(): TaskId = TaskId(UUID.randomUUID().toString, ConsumerGroup)
 
-  var n = 0
+  private[this] var n = 0
 
   override def preStart(): Unit =
     scheduler.scheduleOnce(5.seconds, self, Tick)
@@ -39,10 +37,10 @@ class Producer(masterProxy: ActorRef) extends Actor with ActorLogging {
   }
 
   def waitAccepted(work: Task): Actor.Receive = {
-    case p2c.Accepted(id) =>
+    case p2c.Accepted(_) =>
       context.unbecome()
       scheduler.scheduleOnce(rnd.nextInt(3, 10).seconds, self, Tick)
-    case p2c.Rejected(id) =>
+    case p2c.Rejected(_) =>
       log.info("Work not accepted, retry after a while")
       scheduler.scheduleOnce(3.seconds, masterProxy, work)
   }
